Working on the separation of the declarative aspect of the query from it's execution. Ideally, I would like to propose them as separate project, non binded to a specific system.
However, the construction of the SDS directly depends on the query, which opens a new scenario that requires to think
about what happens in SPARQL.

Moreover, considering stream reasoners on the web, how we assing uris to resources matters. Should they align with their web representation?
How should we model the RSP engine as agent?

I solved the issue temporarily by injecting the resolver into the parser. This however won't be possible if we move to an independent RDF library as RDF lib.