[Core]
I made some main classes generic to simplify the alignment with an external tool for query answering.
I am a bit confused on the role of Commons RDF, Maybe I am not using it as I should.

The current guess is that, if I implement the core/simple (next commit) according only with the interface,
one could implement his own binding to those interfacen and get the job done.

ContinuousQueryExecution is now subclass of observer, Lorenzo is right, we need to move to our implementation.